fix(#2819): core seam — exclude recv-wait from stream_merge, instrument BTI fallback, saturating add
- B2: thread-local recv-wait accumulator (add_pull_wait_nanos/pull_wait_nanos)
  fed by the merge-input blocking recv (refill_buffer), so the drive loop can
  subtract producer-starvation wait from stream_merge (merge CPU only).
- B4: wrap read_next_block (cold_fault) + decompress (decompress) in the
  full-ring fallback stream_all_partitions_for_compaction, so BTI/da scans
  emit all 5 sub-phases (not 3).
- sink_active() Cell<bool> fast gate; add_nanos now saturating via fetch_update
  (correct under concurrent producer writers).
- Fix stale docs (producer thread not feed thread; explicit propagation).
- Drop redundant Histogram count cast (keep ExponentialHistogram's usize cast).

+++ cqlite-core/src/observability/testing.rs
@@ -466,7 +466,7 @@ macro_rules! collect_points_impl {
                             // For histograms, expose the running sum so a test can
                             // assert "something was recorded" and read total magnitude.
                             value: to_f64(dp.sum()),
-                            count: dp.count() as u64,
+                            count: dp.count(),
                             attributes: attrs(dp.attributes()),
                         });
                     }
@@ -474,6 +474,8 @@ macro_rules! collect_points_impl {
                 MetricData::ExponentialHistogram(hist) => {
                     for dp in hist.data_points() {
                         out.push(MetricPoint {
+                            // ExponentialHistogram's `count()` is `usize` (unlike
+                            // Histogram's `u64`), so this cast is load-bearing.
                             value: to_f64(dp.sum()),
                             count: dp.count() as u64,
                             attributes: attrs(dp.attributes()),
